--- src/main/java/com/vishal/wacaj/model/Body.java
@@ -1,5 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Body {
-    String text;
-}
