--- src/main/java/com/vishal/wacaj/model/Button.java
@@ -1,7 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Button {
-    String type;
-    String title;
-    String id;
-}
