--- src/main/java/com/vishal/wacaj/model/Currency.java
@@ -1,7 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Currency {
-    String fallbackValue;
-    String code;
-    float amount1000;
-}
