--- src/main/java/com/vishal/wacaj/model/DateTime.java
@@ -1,5 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class DateTime {
-
-}
