--- src/main/java/com/vishal/wacaj/model/Emails.java
@@ -1,6 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Emails {
-    String email;
-    String type;
-}
