--- src/main/java/com/vishal/wacaj/model/Footer.java
@@ -1,5 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Footer {
-    String text;
-}
