--- src/main/java/com/vishal/wacaj/model/Language.java
@@ -1,6 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Language {
-    String policy;
-    String code;
-}
