--- src/main/java/com/vishal/wacaj/model/Org.java
@@ -1,7 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Org {
-    String company;
-    String department;
-    String title;
-}
