--- src/main/java/com/vishal/wacaj/model/Phones.java
@@ -1,7 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Phones {
-    String phone;
-    String type;
-    String waId;
-}
