--- src/main/java/com/vishal/wacaj/model/Reaction.java
@@ -1,6 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Reaction {
-    String messageId;
-    String emoji;
-}
