--- src/main/java/com/vishal/wacaj/model/Rows.java
@@ -1,7 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Rows {
-    String id;
-    String title;
-    String description;
-}
