--- src/main/java/com/vishal/wacaj/model/Text.java
@@ -1,6 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Text {
-    String body;
-    String previewUrl;
-}
