--- src/main/java/com/vishal/wacaj/model/Urls.java
@@ -1,6 +1,0 @@
-package com.vishal.wacaj.model;
-
-public class Urls {
-    String url;
-    String type;
-}
