+++ src/main/java/com/vishal/wacaj/model/message/Button.java
@@ -1,0 +1,12 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Button {
+    String type;
+    String title;
+    String id;
+}
