+++ src/main/java/com/vishal/wacaj/model/message/Currency.java
@@ -1,0 +1,12 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Currency {
+    String fallbackValue;
+    String code;
+    float amount1000;
+}
