+++ src/main/java/com/vishal/wacaj/model/message/DateTime.java
@@ -1,0 +1,10 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class DateTime {
+
+}
