+++ src/main/java/com/vishal/wacaj/model/message/Interactive.java
@@ -1,5 +1,10 @@
-package com.vishal.wacaj.model;
+package com.vishal.wacaj.model.message;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
 public class Interactive {
     Action action;
     Body body;
