+++ src/main/java/com/vishal/wacaj/model/message/Language.java
@@ -1,0 +1,11 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Language {
+    String policy;
+    String code;
+}
