+++ src/main/java/com/vishal/wacaj/model/message/Location.java
@@ -1,0 +1,13 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Location {
+    String latitude;
+    String longitude;
+    String name;
+    String address;
+}
