+++ src/main/java/com/vishal/wacaj/model/message/Message.java
@@ -1,0 +1,27 @@
+package com.vishal.wacaj.model.message;
+
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Message {
+    Media audio;
+    List<Contact> contacts=new ArrayList<>();
+    Context context;
+    Media document;
+    Media image;
+    Interactive interactive;
+    Location location;
+    String messagingProduct;
+    boolean previewUrl;
+    String recepientType;
+    String status;
+    Media sticker;
+    Template template;
+    Text text;
+    String to;
+    String type;
+}
