+++ src/main/java/com/vishal/wacaj/model/message/Name.java
@@ -1,5 +1,10 @@
-package com.vishal.wacaj.model;
+package com.vishal.wacaj.model.message;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
 public class Name {
     String formattedName;
     String firstName;
