+++ src/main/java/com/vishal/wacaj/model/message/Org.java
@@ -1,0 +1,12 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Org {
+    String company;
+    String department;
+    String title;
+}
