+++ src/main/java/com/vishal/wacaj/model/message/Phones.java
@@ -1,0 +1,12 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Phones {
+    String phone;
+    String type;
+    String waId;
+}
