+++ src/main/java/com/vishal/wacaj/model/message/Product.java
@@ -1,0 +1,9 @@
+package com.vishal.wacaj.model.message;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Product {
+    String productRetailerId;
+}
