+++ src/main/java/com/vishal/wacaj/model/message/Reaction.java
@@ -1,0 +1,11 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Reaction {
+    String messageId;
+    String emoji;
+}
