+++ src/main/java/com/vishal/wacaj/model/message/Rows.java
@@ -1,0 +1,12 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Rows {
+    String id;
+    String title;
+    String description;
+}
