+++ src/main/java/com/vishal/wacaj/model/message/Template.java
@@ -1,8 +1,12 @@
-package com.vishal.wacaj.model;
+package com.vishal.wacaj.model.message;
 
 import java.util.ArrayList;
 import java.util.List;
+import lombok.Getter;
+import lombok.Setter;
 
+@Getter
+@Setter
 public class Template {
     String name;
     Language language;
