+++ src/main/java/com/vishal/wacaj/model/message/Text.java
@@ -1,0 +1,11 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Text {
+    String body;
+    String previewUrl;
+}
