+++ src/main/java/com/vishal/wacaj/model/message/Urls.java
@@ -1,0 +1,11 @@
+package com.vishal.wacaj.model.message;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Urls {
+    String url;
+    String type;
+}
